# api-spec-json.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alpaca API Service",
    "description": "A lightweight API service for Alpaca Markets integration with artificial order management and centralized configuration",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/alpaca-service"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/orders": {
      "post": {
        "summary": "Place order",
        "tags": ["Orders"],
        "operationId": "placeOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit or cooldown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get orders",
        "tags": ["Orders"],
        "operationId": "getOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["open", "closed", "all"]
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "side",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["buy", "sell"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "summary": "Get order details",
        "tags": ["Orders"],
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel order",
        "tags": ["Orders"],
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelOrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/artificial-orders": {
      "post": {
        "summary": "Create artificial order",
        "tags": ["Artificial Orders"],
        "operationId": "createArtificialOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtificialOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artificial order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtificialOrderResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get artificial orders",
        "tags": ["Artificial Orders"],
        "operationId": "getArtificialOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "triggered", "executed", "cancelled"]
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artificial orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtificialOrdersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/artificial-orders/{artificialId}": {
      "get": {
        "summary": "Get artificial order",
        "tags": ["Artificial Orders"],
        "operationId": "getArtificialOrder",
        "parameters": [
          {
            "name": "artificialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artificial order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtificialOrderDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artificial order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel artificial order",
        "tags": ["Artificial Orders"],
        "operationId": "cancelArtificialOrder",
        "parameters": [
          {
            "name": "artificialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artificial order cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelArtificialOrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artificial order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/market/quote/{symbol}": {
      "get": {
        "summary": "Get latest quote",
        "tags": ["Market Data"],
        "operationId": "getQuote",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/bars/{symbol}": {
      "get": {
        "summary": "Get historical bars",
        "tags": ["Market Data"],
        "operationId": "getBars",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1Min", "5Min", "15Min", "1Hour", "1Day"]
            }
          },
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical bars",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderRequest": {
        "type": "object",
        "required": ["symbol", "qty", "side", "orderType", "timeInForce"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Stock symbol"
          },
          "qty": {
            "type": "number",
            "description": "Order quantity"
          },
          "side": {
            "type": "string",
            "enum": ["buy", "sell"],
            "description": "Order side"
          },
          "orderType": {
            "type": "string",
            "enum": ["market", "limit", "stop", "stop_limit"],
            "description": "Order type"
          },
          "timeInForce": {
            "type": "string",
            "enum": ["day", "gtc", "opg", "cls", "ioc", "fok"],
            "description": "Time in force"
          },
          "limitPrice": {
            "type": "number",
            "description": "Limit price (required for limit orders)"
          },
          "stopPrice": {
            "type": "number",
            "description": "Stop price (required for stop orders)"
          },
          "clientOrderId": {
            "type": "string",
            "description": "Client-specified order ID"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "clientOrderId": { "type": "string" },
          "symbol": { "type": "string" },
          "qty": { "type": "number" },
          "filledQty": { "type": "number" },
          "side": { "type": "string" },
          "orderType": { "type": "string", "readOnly": true, "description": "Order type derived by the service based on provided parameters" },
          "timeInForce": { "type": "string" },
          "limitPrice": { "type": "number" },
          "stopPrice": { "type": "number" },
          "status": { "type": "string" },
          "filledAvgPrice": { "type": "number" },
          "duplicateCheckPassed": { "type": "boolean" },
          "rateLimitStatus": { "type": "string" },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "submittedAt": { "type": "string", "format": "date-time" },
          "filledAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ArtificialOrderRequest": {
        "type": "object",
        "required": ["symbol", "qty", "side", "triggerPrice"],
        "properties": {
          "symbol": { "type": "string" },
          "qty": { "type": "number" },
          "side": { "type": "string", "enum": ["buy", "sell"] },
          "triggerPrice": { "type": "number" },
          "limitPrice": { "type": "number" },
          "timeInForce": { "type": "string", "enum": ["day", "gtc"] },
          "notes": { "type": "string" }
        }
      },
      "ArtificialOrder": {
        "type": "object",
        "properties": {
          "artificialId": { "type": "string" },
          "symbol": { "type": "string" },
          "qty": { "type": "number" },
          "side": { "type": "string" },
          "triggerPrice": { "type": "number" },
          "limitPrice": { "type": "number" },
          "orderType": { "type": "string", "readOnly": true, "description": "Order type derived by the service based on provided parameters" },
          "status": { "type": "string" },
          "currentPrice": { "type": "number" },
          "distanceFromTrigger": { "type": "number" },
          "distancePercent": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "triggeredAt": { "type": "string", "format": "date-time" },
          "executedOrderId": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "detail": { "type": "string" },
              "fields": { "type": "object", "description": "Map of field names to error messages for validation errors" },
              "metadata": { "type": "object" }
            }
          }
        }
      },
...
      "OrderResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/Order" }
            }
          }
        ]
      },
      "OrdersResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "orders": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Order" }
                  },
                  "nextCursor": { "type": "string" }
                }
              }
            }
          }
        ]
      },
      "OrderDetailResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "allOf": [
                  { "$ref": "#/components/schemas/Order" },
                  {
                    "type": "object",
                    "properties": {
                      "remainingQty": { "type": "number" },
                      "fills": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "fillId": { "type": "string" },
                            "qty": { "type": "number" },
                            "price": { "type": "number" },
                            "filledAt": { "type": "string", "format": "date-time" }
                          }
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
...
      "CancelOrderResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "orderId": { "type": "string" },
                  "status": { "type": "string" },
                  "canceledAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        ]
      },
      "ArtificialOrderResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/ArtificialOrder" }
            }
          }
        ]
      },
      "ArtificialOrdersResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ArtificialOrder" }
              }
            }
          }
        ]
      },
      "ArtificialOrderDetailResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/ArtificialOrder" }
            }
          }
        ]
      },
      "CancelArtificialOrderResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "artificialId": { "type": "string" },
                  "status": { "type": "string" },
                  "cancelledAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        ]
      },
...
      "QuoteResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string" },
                  "bidPrice": { "type": "number" },
                  "bidSize": { "type": "integer" },
                  "askPrice": { "type": "number" },
                  "askSize": { "type": "integer" },
                  "lastPrice": { "type": "number" },
                  "lastSize": { "type": "integer" },
                  "timestamp": { "type": "string", "format": "date-time" },
                  "conditions": {
                    "type": "array",
                    "items": { "type": "string" }
                  },
                  "tape": { "type": "string" }
                }
              }
            }
          }
        ]
      },
      "Bar": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "open": { "type": "number" },
          "high": { "type": "number" },
          "low": { "type": "number" },
          "close": { "type": "number" },
          "volume": { "type": "integer" },
          "vwap": { "type": "number" },
          "tradeCount": { "type": "integer" }
        }
      },
      "BarsResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string" },
                  "bars": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Bar" }
                  },
                  "nextPageToken": { "type": "string" }
                }
              }
            }
          }
        ]
      },
...
    }
  },
...
}